% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareQueriesByState.R
\name{compareQueriesByState}
\alias{compareQueriesByState}
\title{Takes two queries from the user and returns the query with more results by state}
\usage{
compareQueriesByState(query1, query2)
}
\arguments{
\item{query1}{First restaurant-related query}

\item{query2}{First restaurant-related query}
}
\description{
Takes two queries from the user and returns the query with more results by state
}
\examples{
compareQueriesByState("Pizza Hut","Dominoe's")
}

