% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlcoholFriendlyRestaurants.R
\name{getAlcoholFriendlyRestaurants}
\alias{getAlcoholFriendlyRestaurants}
\title{Given a city name and a cusisine type, provide a list of restaurants that serve alcohol}
\usage{
getAlcoholFriendlyRestaurants(city, cuisine, full = FALSE)
}
\arguments{
\item{city}{A city name}

\item{cuisine}{Cuisine name like "Indian", "Thai" etc.}

\item{full}{if TRUE returns full factual response. Default is FALSE,
which results in a smaller data frame with just restaurant name, latitude and longitude}
}
\description{
Given a city name and a cusisine type, provide a list of restaurants that serve alcohol
}
\examples{
getAlcoholFriendlyRestaurants("Mountain View","Indian")
}

