% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQueryCountByState.R
\name{getQueryCountByUSState}
\alias{getQueryCountByUSState}
\title{Gets count of restaurant results matching a query by US state}
\usage{
getQueryCountByUSState(query)
}
\arguments{
\item{query}{a restaurant-related query like "Subway" or "Indian"}
}
\description{
Gets count of restaurant results matching a query by US state
}
\examples{
getQueryCountByUSState("Chipotle")
}

