% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRestaurantByDietType.R
\name{getRestaurantByDietType}
\alias{getRestaurantByDietType}
\title{Given a city name, a cusisine type and a diet type, provide a list of restaurants that meet these conditions}
\usage{
getRestaurantByDietType(city, dietType, cuisine, full)
}
\arguments{
\item{city}{A city name}

\item{dietType}{One of diet types: "vegan","vegetarian","gluttenfree","organic","healthy"}

\item{cuisine}{Cuisine name like "Indian", "Thai" etc.}

\item{full}{if TRUE returns full factual response. Default is FALSE,
which results in a smaller data frame with just restaurant name, latitude and longitude}
}
\description{
Given a city name, a cusisine type and a diet type, provide a list of restaurants that meet these conditions
}
\examples{
getKidsFriendlyRestaurants("New York","Thai")
}

