% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRestaurantByPriceRange.R
\name{getRestaurantByPriceRange}
\alias{getRestaurantByPriceRange}
\title{Given a city name, a cusisine type and a price-range type, provide a list of restaurants that meet these conditions}
\usage{
getRestaurantByPriceRange(city, priceRange, cuisine, full = FALSE)
}
\arguments{
\item{city}{A city name}

\item{priceRange}{One of these: "very inexpensive","inexpensive","moderate","expensive","very expensive"}

\item{cuisine}{Cuisine name like "Indian", "Thai" etc.}

\item{full}{if TRUE returns full factual response. Default is FALSE,
which results in a smaller data frame with just restaurant name, latitude and longitude}
}
\description{
Given a city name, a cusisine type and a price-range type, provide a list of restaurants that meet these conditions
}
\examples{
getRestaurantByPriceRange("San Francisco","moderate","Asian")
}

