% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRestaurantsAcceptingReservations.R
\name{getRestaurantsAcceptingReservations}
\alias{getRestaurantsAcceptingReservations}
\title{Given a city name and a cusisine type, provide a list of restaurants that accept reservations}
\usage{
getRestaurantsAcceptingReservations(city, cuisine, full = FALSE)
}
\arguments{
\item{city}{A city name}

\item{cuisine}{Cuisine name like "Indian", "Thai" etc.}

\item{full}{if TRUE returns full factual response. Default is FALSE,
which results in a smaller data frame with just restaurant name, latitude and longitude}
}
\description{
Given a city name and a cusisine type, provide a list of restaurants that accept reservations
}
\examples{
getRestaurantsAcceptingReservations("San Francisco","Asian")
}

