% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRestaurantsByMealType.R
\name{getRestaurantsByMealType}
\alias{getRestaurantsByMealType}
\title{Given a city name, a cusisine type and a meal type, provide a list of restaurants that meet these conditions}
\usage{
getRestaurantsByMealType(city, mealType, cuisine, full = FALSE)
}
\arguments{
\item{city}{A city name}

\item{mealType}{One of these: "lunch","dinner","breakfast"}

\item{cuisine}{Cuisine name like "Indian", "Thai" etc.}

\item{full}{if TRUE returns full factual response. Default is FALSE,
which results in a smaller data frame with just restaurant name, latitude and longitude}
}
\description{
Given a city name, a cusisine type and a meal type, provide a list of restaurants that meet these conditions
}
\examples{
getRestaurantsByMealType("Bellevue","Lunch","Thai")
}

