% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRestaurantsNearLocation.R
\name{getRestaurantsNearLocation}
\alias{getRestaurantsNearLocation}
\title{Given a point ofinterest, provide a list of nearby restaurants}
\usage{
getRestaurantsNearLocation(pointName, category)
}
\arguments{
\item{pointName}{A point of interest like "Golden Gate Bridge" or "Empire State Building"}

\item{category}{Category name}
}
\description{
Given a point ofinterest, provide a list of nearby restaurants
}
\examples{
getRestaurantsNearLocation("Space Needle","Asian")
}

