% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRestaurantsWithDeliveryOrTakeOut.R
\name{getRestaurantsWithDeliveryOrTakeOut}
\alias{getRestaurantsWithDeliveryOrTakeOut}
\title{Given a city name, a cusisine type and an order type (takeout or delivery),
provide a list of restaurants that meet these conditions}
\usage{
getRestaurantsWithDeliveryOrTakeOut(city, orderType, cuisine, full = FALSE)
}
\arguments{
\item{city}{A city name}

\item{orderType}{"delivery","takeout"}

\item{cuisine}{Cuisine name like "Indian", "Thai" etc.}

\item{full}{if TRUE returns full factual response. Default is FALSE,
which results in a smaller data frame with just restaurant name, latitude and longitude}
}
\description{
Given a city name, a cusisine type and an order type (takeout or delivery),
provide a list of restaurants that meet these conditions
}
\examples{
getRestaurantsWithDeliveryOrTakeOut("San Francisco","delivery","Indian")
}

