% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWheelChairAccesibleRestaurants.R
\name{getWheelChairAccesibleRestaurants}
\alias{getWheelChairAccesibleRestaurants}
\title{Given a city name and a cusisine type, provide a list of restaurants that are wheelchair accessible}
\usage{
getWheelChairAccesibleRestaurants(city, cuisine, full = FALSE)
}
\arguments{
\item{city}{A city name}

\item{cuisine}{Cuisine name like "Indian", "Thai" etc.}

\item{full}{if TRUE returns full factual response. Default is FALSE,
which results in a smaller data frame with just restaurant name, latitude and longitude}
}
\description{
Given a city name and a cusisine type, provide a list of restaurants that are wheelchair accessible
}
\examples{
getWheelChairAccesibleRestaurants("San Francisco","Asian")
}

