% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeFactual.R
\name{initializeFactual}
\alias{initializeFactual}
\title{Initialize factual and set factualAPIKey as a global variable
where it can be accessed by other functions}
\usage{
initializeFactual(key)
}
\arguments{
\item{key}{a restaurant-related query like "Subway" or "Indian"}
}
\description{
Initialize factual and set factualAPIKey as a global variable
where it can be accessed by other functions
}
\examples{
initializeFactual("mKxC6I9lTWnKNTSNF12e3keaWblCXqoaZ1qROdVo")
}

