% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapQueryCountByUSState.R
\name{mapQueryCountByUSState}
\alias{mapQueryCountByUSState}
\title{Takes two queries from the user and returns the query with more results by state}
\usage{
mapQueryCountByUSState(query)
}
\arguments{
\item{query}{query a restaurant-related query like "Subway" or "Indian"}
}
\description{
Takes two queries from the user and returns the query with more results by state
}
\examples{
mapQueryCountByUSState ("Chipotle")
}

