% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NYC_business_casual.R
\docType{data}
\name{NYC_business_casual}
\alias{NYC_business_casual}
\title{NYC_business_casual data}
\description{
Test data containing  restaurants with attire type "business casual" in New York
}
\keyword{data}

