% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yelpR.R
\docType{class}
\name{yelpR}
\alias{yelpR}
\title{Create a \code{yelpR} object}
\format{An \code{\link{R6Class}} generator object}
\usage{
yelpR
}
\description{
\code{yelpR} is a restaurant object that encapsulates the data returned from yelp.
A yelp restaurant data object consisists of 2 main components, connection settings and data settings. connection settings
include the keys that are needed to authenticate with yelp. The data object stores the restaurant results that match a
specific criteria.
}
\section{Fields}{

\describe{
\item{\code{consumerKey}}{consumerKey used to establish a yelp connection}

\item{\code{consumerSecret}}{consumerSecret used to establish a yelp connection}

\item{\code{token}}{token used to establish a yelp connection}

\item{\code{token_secret}}{token_secret used to establish a yelp connection}

\item{\code{data}}{The most recent restaurant data set that queried by the user displaying top restaurants to analyze at the top. Two type of queries are supported - restaurant data
location and cusine type or
location, cusine and deals avaialble
We apply a sorting algorithm to sort the results based on both the review and rating of each restaurant
taking into account both the review and rating of each restaurant. Thereby the data from there by the user. Typically the data set contains 27
columns defined by yelp that describe information about the restaurant viz.
"region.span.latitude_delta"      "region.span.longitude_delta"     "region.center.latitude"
"region.center.longitude"         "total"                           "businesses.is_claimed"
"businesses.rating"               "businesses.mobile_url"           "businesses.rating_img_url"
"businesses.review_count"         "businesses.name"                 "businesses.rating_img_url_small"
"businesses.url"                  "businesses.categories"           "businesses.phone"
"businesses.snippet_text"         "businesses.image_url"            "businesses.snippet_image_url"
"businesses.display_phone"        "businesses.rating_img_url_large" "businesses.id"
"businesses.is_closed"            "businesses.menu_date_updated"    "businesses.menu_provider"
"dup.row.names"                   "businesses.gift_certificates"    "businesses.deals"
The user then queries this data frame for further results depending on the nature of the task}
}}
\section{Methods}{


\describe{
  \item{\code{yelpR$new()}}{Creates a empty \code{yelpR}
        object which is then initialized}
  \item{\code{InitializeYelp(consumerKey,consumerSecret,token,token_secret)}}{Authenticates the user to establish a yelp connection.
         Returns true if the conntection was successfully established. otherwise displays an error with return value NA. Warning
         messages are displayed with error value NULL.}
  \item{\code{getData()}}{Returns the restaurant data of the last issued query.}
  \item{\code{queryData(city, state, cusine)}}{Returns restaurant data as a data frame based on the filter criteria specified.
               InitializeYelp should return true for this method to be successful.}
   \item{\code{shortlistRestaurants(city, state, cusine)}}{Returns 6 shortlisted restaurants as a data frame based on the filter criteria specified.
               InitializeYelp should return true for this method to be successful. Basic information about the restaurant such as name, ph, url,
               rating and review are avaialble for the user to analyze.}
   \item{\code{getRestaurantWithDeals(city, state, cusine)}}{Returns restaurants that offer deals based on the filter criteria specified.
               InitializeYelp should return true for this method to be successful. All information about the restaurant is returned for the usser
               to analyze }

}
}
\examples{
restaurantObj <- yelpR$new()
restaurantObj$InitializeYelp()
restaurantDF = restaurantObj$queryData("Stanford","CA","Indian")
head(restaurantDF)$businesses.name

restaurantDF = restaurantObj$queryData("San Francisco","CA","Spanish")
head(restaurantDF)$businesses.name

restaurantDF = restaurantObj$shortlistRestaurants("Napa Valley","CA","Italian")
print(restaurantDF)
print("Take your significant other, goto these places and Enjoy !")

restaurantDF = restaurantObj$getRestaurantWithDeals("San Francisco","CA","Indian")
head(restaurantDF)$businesses.name


}
\keyword{datasets}

