% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MountainView_Indian_Alcohol_Friendly.R
\docType{data}
\name{MountainView_Indian_Alcohol_Friendly}
\alias{MountainView_Indian_Alcohol_Friendly}
\title{MountainView_Indian_Alcohol_Friendly data}
\description{
Test data containing  Indian Alcoholo friendly restaurants in Mountain View, CA
}
\keyword{data}

