% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBusinesses.R
\name{plotBusinesses}
\alias{plotBusinesses}
\title{Plot business on a map from a city name and a date-frame of business name, longitude and latitude}
\usage{
plotBusinesses(city, businessDF)
}
\arguments{
\item{city}{A city name}

\item{businessDF}{a date-frame of business name, longitude and latitude}
}
\description{
Plot business on a map from a city name and a date-frame of business name, longitude and latitude
}
\examples{
plotBusinesses("New York",NYC_Vegan)
}

