% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRestaurantsNearLocation.R
\name{plotRestaurantsNearLocation}
\alias{plotRestaurantsNearLocation}
\title{Given a point ofinterest, provide a list of nearby restaurants}
\usage{
plotRestaurantsNearLocation(pointName, cuisine)
}
\arguments{
\item{pointName}{A point of interest like "Golden Gate Bridge" or "Empire State Building"}

\item{cuisine}{Cuisine name like "Indian", "Thai" etc.}
}
\description{
Given a point ofinterest, provide a list of nearby restaurants
}
\examples{
plotRestaurantsNearLocation("Empire State Building", "Pizza")
}

